Moved backend files into the backend folder on plaid branch
